+++ packages/client/src/app/challenge/components/ChallengeCheckBox.vue
